Refactor TabmanBar.BackgroundView
