+++ Sources/Tabman/TabmanBar/TabmanBar+Appearance.swift
@@ -72,7 +72,7 @@ public extension TabmanBar {
         
         public struct Style {
             /// The background style for the bar.
-            public var background: TabmanBarBackgroundView.BackgroundStyle?
+            public var background: BackgroundView.BackgroundStyle?
             /// Whether to show a fade on the items at the bounds edge of a scrolling button bar.
             public var showEdgeFade: Bool?
             /// Color of the separator at the bottom of the bar.
