+++ Sources/Tabman/TabmanBar/TabmanBar.swift
@@ -97,7 +97,7 @@ open class TabmanBar: UIView, TabmanBarLifecycle {
     }
     
     /// Background view of the bar.
-    public private(set) var backgroundView: TabmanBarBackgroundView = TabmanBarBackgroundView(forAutoLayout: ())
+    public private(set) var backgroundView: BackgroundView = BackgroundView(forAutoLayout: ())
     /// The content view for the bar.
     public private(set) var contentView = UIView(forAutoLayout: ())
     /// The bottom separator view for the bar.
